IGAPP-400: Move imprint to avoid being copied by webpack
